opus/requirements.txt
oauthlib==3.2.2
Requests==2.31.0
requests_oauthlib==1.3.1
websocket_client==1.6.4
